requirements.txt
selenium=4.20.0
behave=1.2.6
Appium-Python-Client=4.0.0
allure-behave=2.13.5